requirements.txt
django
djangorestframework
gunicorn